refactoring: move "slow" test decorator to test.utils
