+++ tests/inside_worker_test/nonop_way_test.py
@@ -6,7 +6,7 @@ from sqlalchemy.sql.schema import Table as DbTable
 
 from osmaxx.utils.frozendict import frozendict
 from tests.conftest import TagCombination
-from tests.inside_worker_test.conftest import slow
+from tests.utils import slow
 from tests.inside_worker_test.declarative_schema import osm_models
 
 MAJOR_KEYS = frozenset({'highway', 'railway'})
