+++ tests/utils/__init__.py
@@ -1,0 +1,6 @@
+import pytest
+
+slow = pytest.mark.skipif(
+    not pytest.config.getoption("--runslow"),
+    reason="need --runslow option to run"
+)
